pyproject.toml
[project]
name = "rss-scrape"
version = "0.1.0"
description = "RSS/Atom scraper service with MariaDB storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "SQLAlchemy==2.0.36",
    "feedparser==6.0.11",
    "httpx==0.27.2",
    "pymysql==1.1.1",
    "python-dateutil==2.9.0.post0",
]
